Restructure & split sidebar component and part of toolbar
