+++ apps/web-mzima-client/src/app/app.component.html
@@ -3,7 +3,8 @@
 <mat-sidenav-container
   [dir]="isRTL ? 'rtl' : 'ltr'"
   class="main"
-  [ngClass]="{ 'main--inner': !(isDesktop$ | async)! && isInnerPage }">
+  [ngClass]="{ 'main--inner': !(isDesktop$ | async)! && isInnerPage }"
+>
   <mat-sidenav opened mode="side" disableClose class="sidebar">
     <app-sidebar></app-sidebar>
   </mat-sidenav>
@@ -14,6 +15,16 @@
     ></app-toolbar>
     <div class="content-wrapper">
       <router-outlet></router-outlet>
+    </div>
+    <div class="mobile-page-nav" *ngIf="!isInnerPage">
+      <app-submit-post-button
+        class="mobile-page-nav__add-post-btn"
+        [isRound]="true"
+        [hasIconOnly]="true"
+        [hasText]="false"
+      >
+      </app-submit-post-button>
+      <app-menu-list-links></app-menu-list-links>
     </div>
   </mat-sidenav-content>
 </mat-sidenav-container>
