+++ apps/web-mzima-client/src/app/app.component.scss
@@ -4,6 +4,7 @@
   body:not(.burger-menu-open) {
     .main:not(.main--inner) {
       @include breakpoint-max($tablet) {
+        // Adds padding bottom for smaller devices so that page links at the bottom does not cover some content (e.g. in activity & settings)
         padding-bottom: 80px;
       }
     }
@@ -37,16 +38,48 @@
     max-width: 200px;
   }
 
+  // Hide sidebar & content on smaller devices
+  // Reminder: May or may not need this (if the ngIf directive hides mat-sidenav)
   @include breakpoint-max($tablet) {
-    right: 0;
-    top: auto;
-    position: fixed;
-    max-width: none;
-    overflow: visible;
-    z-index: 500 !important;
+    display: none;
   }
 }
 
+.mobile-page-nav {
+  display: none;
+  position: fixed;
+  width: 100%;
+  bottom: 0;
+  left: 0;
+  max-width: none;
+  overflow: visible;
+  z-index: 500 !important;
+  margin: 0;
+  opacity: 1;
+
+  @include breakpoint-max($tablet) {
+    display: block;
+  }
+
+  &__add-post-btn {
+    @include breakpoint-max($laptop-small) {
+      padding-left: 16px;
+      padding-right: 16px;
+    }
+
+    @include breakpoint-max($tablet) {
+      left: 50%;
+      padding: 0;
+      z-index: 25;
+      bottom: 100%;
+      border: none;
+      position: absolute;
+      margin: 0 0 -12px -32px;
+    }
+  }
+}
+
+// Needed for opening mobile-menu
 ::ng-deep {
   body.burger-menu-open {
     .sidebar {
