+++ apps/web-mzima-client/src/app/shared/components/fragments/menu-list-links/menu-list-links.component.scss
@@ -1,81 +1,13 @@
 @import 'helpers';
 
 .sidebar {
-  height: 100%;
-  display: flex;
-  font-size: 12px;
-  overflow-y: auto;
-  overflow-x: hidden;
-  line-height: 14px;
-  position: relative;
-  --side-offset: 24px;
-  flex-direction: column;
-  scroll-behavior: smooth;
-  color: var(--color-neutral-50);
-  -webkit-overflow-scrolling: touch;
-
-  @include breakpoint-min($tablet) {
-    padding: 24px var(--side-offset) 20px;
-  }
-
-  @include breakpoint-max($tablet) {
-    overflow: visible;
-  }
-
-  &__logo {
-    height: 72px;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    justify-content: center;
-    padding: 8px var(--side-offset) 0;
-    border-bottom: 1px solid var(--color-neutral-90);
-    margin: -24px calc(var(--side-offset) * -1) 8px;
-
-    img {
-      display: block;
-      max-width: 100%;
-    }
-  }
-
-  &__add-post-btn {
-    display: block;
-    border-width: 1px 0;
-    border-style: solid;
-    padding: 12px var(--side-offset);
-    border-color: var(--color-neutral-90);
-    margin: 0 calc(var(--side-offset) * -1) 8px;
-
-    @include breakpoint-max($laptop-small) {
-      padding-left: 16px;
-      padding-right: 16px;
-    }
-
-    @include breakpoint-max($tablet) {
-      left: 50%;
-      padding: 0;
-      z-index: 25;
-      bottom: 100%;
-      border: none;
-      position: absolute;
-      margin: 0 0 -12px -32px;
-    }
-  }
-
   &__menu {
-    @include breakpoint-min($tablet) {
-      margin: 0 calc(var(--side-offset) * -1);
-
-      &:not(:first-child) {
-        margin-top: 8px;
-        padding-top: 8px;
-        border-top: 1px solid var(--color-neutral-90);
-      }
-    }
+    margin: 0 calc(var(--side-offset) * -1);
+    color: var(--color-neutral-50);
 
     @include breakpoint-max($tablet) {
       display: flex;
-      justify-content: flex-start;
+      justify-content: stretch;
     }
   }
 }
