+++ apps/web-mzima-client/src/app/shared/components/fragments/menu-list-non-links/menu-list-non-links.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MenuListNonLinksComponent } from './menu-list-non-links.component';
+
+describe('MenuListNonLinksComponent', () => {
+  let component: MenuListNonLinksComponent;
+  let fixture: ComponentFixture<MenuListNonLinksComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [MenuListNonLinksComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(MenuListNonLinksComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
