+++ apps/web-mzima-client/src/app/shared/components/fragments/mobile-menu/mobile-menu.component.html
@@ -1,0 +1,6 @@
+<!-- Smaller screens only -->
+<div class="menu" *ngIf="!(isDesktop$ | async)!" [ngClass]="{ 'menu--open': isBurgerMenuOpen }">
+  <app-deployment-details></app-deployment-details>
+  <app-menu-list-non-links></app-menu-list-non-links>
+  <app-share-and-donate></app-share-and-donate>
+</div>
