+++ apps/web-mzima-client/src/app/shared/components/fragments/mobile-menu/mobile-menu.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MobileMenuComponent } from './mobile-menu.component';
+
+describe('MobileMenuComponent', () => {
+  let component: MobileMenuComponent;
+  let fixture: ComponentFixture<MobileMenuComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [MobileMenuComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(MobileMenuComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
