+++ apps/web-mzima-client/src/app/shared/components/fragments/share-and-donate/share-and-donate.component.html
@@ -1,0 +1,22 @@
+<!-- Desktop -->
+<div class="toolbar__controls" *ngIf="isDesktop$ | async">
+  <app-donation-button class="toolbar__button" *ngIf="isDonateAvailable"></app-donation-button>
+  <mzima-client-button (buttonClick)="openShare()" color="secondary" class="toolbar__button">
+    <span>{{ 'app.share' | translate }}</span>
+    <mat-icon icon svgIcon="share"></mat-icon>
+  </mzima-client-button>
+</div>
+
+<!-- Smaller screens -->
+<div class="menu__footer" *ngIf="!(isDesktop$ | async)!">
+  <app-donation-button class="menu__footer__button" *ngIf="isDonateAvailable"></app-donation-button>
+  <mzima-client-button
+    fill="outline"
+    color="custom"
+    (buttonClick)="openShare()"
+    class="menu__footer__button"
+  >
+    {{ 'app.share' | translate }}
+    <mat-icon icon svgIcon="share"></mat-icon>
+  </mzima-client-button>
+</div>
