+++ apps/web-mzima-client/src/app/shared/components/fragments/share-and-donate/share-and-donate.component.scss
@@ -1,0 +1,37 @@
+@import 'helpers';
+
+.toolbar {
+  &__controls {
+    display: flex;
+
+    @include breakpoint-max($tablet) {
+      display: none;
+    }
+  }
+
+  &__button {
+    margin-left: 5px;
+  }
+}
+
+.menu {
+  &__footer {
+    display: flex;
+    flex-shrink: 0;
+    padding: 16px 12px;
+    background: var(--color-neutral-100);
+
+    &__button {
+      margin: 0 4px;
+      flex: 1 1 calc(50% - 8px);
+      --color: var(--color-light);
+      --background-color: var(--color-black);
+      --border-color: var(--color-neutral-20);
+      --hover-background-color: var(--color-black);
+    }
+
+    @include breakpoint-min($tablet) {
+      display: none;
+    }
+  }
+}
