+++ apps/web-mzima-client/src/app/shared/components/fragments/share-and-donate/share-and-donate.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ShareAndDonateComponent } from './share-and-donate.component';
+
+describe('ShareAndDonateComponent', () => {
+  let component: ShareAndDonateComponent;
+  let fixture: ComponentFixture<ShareAndDonateComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ShareAndDonateComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(ShareAndDonateComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
