+++ apps/web-mzima-client/src/app/shared/components/index.ts
@@ -1,5 +1,10 @@
-export * from './sidebar/sidebar.component';
+export * from './main/sidebar/sidebar.component';
 export * from './toolbar/toolbar.component';
+export * from './fragments/deployment-details/deployment-details.component';
+export * from './fragments/menu-list-non-links/menu-list-non-links.component';
+export * from './fragments/mobile-menu/mobile-menu.component';
+export * from './fragments/menu-list-links/menu-list-links.component';
+export * from './fragments/share-and-donate/share-and-donate.component';
 export * from './submit-post-button/submit-post-button.component';
 export * from './select-languages-modal/select-languages-modal.component';
 export * from './spinner/spinner.component';
