+++ apps/web-mzima-client/src/app/shared/components/main/sidebar/sidebar.component.html
@@ -1,0 +1,16 @@
+<div class="sidebar" *ngIf="isDesktop || !isInnerPage">
+  <app-deployment-details></app-deployment-details>
+  <app-submit-post-button
+    class="sidebar__add-post-btn"
+    [isRound]="false"
+    [hasIconOnly]="false"
+    [hasText]="true"
+  >
+  </app-submit-post-button>
+
+  <nav class="sidebar__controls">
+    <app-menu-list-links></app-menu-list-links>
+    <app-menu-list-non-links></app-menu-list-non-links>
+  </nav>
+  <pre [style]="{ display: 'none' }">>1.0.1</pre>
+</div>
