--- apps/web-mzima-client/src/app/shared/components/sidebar/sidebar.component.html
@@ -1,89 +1,0 @@
-<div class="sidebar" *ngIf="isDesktop || !isInnerPage">
-  <div class="sidebar__main">
-    <span class="sidebar__logo" *ngIf="isDesktop">
-      <img
-        src="../../../../assets/images/ushahidi-logo.svg"
-        alt="ushahidi"
-        width="172"
-        height="40"
-      />
-    </span>
-
-    <div class="sidebar__info">
-      <app-company-info *ngIf="isDesktop"></app-company-info>
-      <app-submit-post-button class="sidebar__add-post-btn"></app-submit-post-button>
-    </div>
-  </div>
-
-  <nav class="sidebar__controls">
-    <div class="sidebar__menu" id="sidebar-menu">
-      <ng-container *ngFor="let item of menu">
-        <ng-container *ngIf="!item.hidden && (!item.adminGuard || isHost)">
-          <a
-            [id]="item.ref"
-            class="sidebar-menu-button"
-            [data-qa]="'btn-' + item.icon"
-            [attr.data-onboard-id]="'sidebar-btn-' + item.icon"
-            *ngIf="item.router; else menuAction"
-            [routerLink]="createRouterLink(item.router)"
-            routerLinkActive="sidebar-menu-button--active"
-            (click)="registerPage($event, item.router, item.label)"
-          >
-            <mat-icon class="sidebar-menu-button__icon" [svgIcon]="item.icon"></mat-icon>
-            <span class="sidebar-menu-button__title">
-              {{ item.label | translate }}
-            </span>
-          </a>
-          <ng-template #menuAction>
-            <button
-              [id]="item.ref"
-              class="sidebar-menu-button"
-              [data-qa]="'btn-' + item.icon"
-              [attr.data-onboard-id]="'sidebar-btn-' + item.icon"
-              (click)="item.action && item.action()"
-            >
-              <mat-icon class="sidebar-menu-button__icon" [svgIcon]="item.icon"></mat-icon>
-              <span class="sidebar-menu-button__title">
-                {{ item.label | translate }}
-              </span>
-            </button>
-          </ng-template>
-        </ng-container>
-      </ng-container>
-    </div>
-    <div class="sidebar__menu" *ngIf="isDesktop">
-      <ng-container *ngFor="let item of userMenu">
-        <ng-container *ngIf="item.visible && (!item.adminGuard || isHost)">
-          <button
-            [id]="item.ref"
-            (click)="item.action()"
-            class="sidebar-menu-button"
-            [data-qa]="'btn-' + item.icon"
-            *ngIf="item.action; else menuLink"
-            [attr.data-onboard-id]="'sidebar-btn-' + item.icon"
-          >
-            <mat-icon class="sidebar-menu-button__icon" [svgIcon]="item.icon"></mat-icon>
-            <span class="sidebar-menu-button__title">{{ item.label | translate }}</span>
-          </button>
-          <ng-template #menuLink>
-            <a
-              [id]="item.ref"
-              class="sidebar-menu-button"
-              [data-qa]="'btn-' + item.icon"
-              (click)="item.action && item.action()"
-              [attr.data-onboard-id]="'sidebar-btn-' + item.icon"
-              routerLinkActive="sidebar-menu-button--active"
-              [routerLink]="item.router && createRouterLink(item.router)"
-            >
-              <mat-icon class="sidebar-menu-button__icon" [svgIcon]="item.icon"></mat-icon>
-              <span class="sidebar-menu-button__title">
-                {{ item.label | translate }}
-              </span>
-            </a>
-          </ng-template>
-        </ng-container>
-      </ng-container>
-    </div>
-  </nav>
-  <pre [style]="{ display: 'none' }">>1.0.1</pre>
-</div>
