+++ apps/web-mzima-client/src/app/shared/components/toolbar/toolbar.component.html
@@ -18,13 +18,7 @@
   </h1>
   <div class="toolbar__controls">
     <app-search-form class="search-form" *ngIf="showSearchForm"></app-search-form>
-    <ng-container *ngIf="isDesktop$ | async">
-      <app-donation-button class="toolbar__button" *ngIf="isDonateAvailable"></app-donation-button>
-      <mzima-client-button (buttonClick)="openShare()" color="secondary" class="toolbar__button">
-        <span>{{ 'app.share' | translate }}</span>
-        <mat-icon icon svgIcon="share"></mat-icon>
-      </mzima-client-button>
-    </ng-container>
+    <app-share-and-donate *ngIf="isDesktop$ | async"></app-share-and-donate>
     <app-language
       class="toolbar__language"
       [languages]="languages"
@@ -65,72 +59,11 @@
   </mzima-client-button>
 </mat-toolbar>
 
-<div class="menu" *ngIf="!(isDesktop$ | async)!" [ngClass]="{ 'menu--open': isBurgerMenuOpen }">
-  <div class="menu__head">
-    <div class="menu__logo">
-      <img
-        src="../../../../assets/images/ushahidi-logo.svg"
-        alt="ushahidi"
-        width="177"
-        height="54"
-      />
-    </div>
-    <mzima-client-button
-      fill="outline"
-      color="custom"
-      [iconOnly]="true"
-      class="menu__head__close"
-      (buttonClick)="toggleBurgerMenu(false)"
-      [ariaLabel]="'modal.button.close' | translate"
-    >
-      <mat-icon icon svgIcon="close"></mat-icon>
-    </mzima-client-button>
-  </div>
-
-  <div class="menu__body">
-    <app-company-info></app-company-info>
-
-    <div class="mobile-menu">
-      <ng-container *ngFor="let item of menu">
-        <hr *ngIf="item.separator" />
-        <a
-          *ngIf="item.visible"
-          class="mobile-menu-button"
-          [data-qa]="'btn-' + item.icon"
-          (click)="item.action && item.action()"
-        >
-          <mat-icon class="mobile-menu-button__icon" [svgIcon]="item.icon"></mat-icon>
-          <span class="mobile-menu-button__title">
-            {{ item.label | translate }}
-          </span>
-        </a>
-      </ng-container>
-      <hr />
-      <a class="mobile-menu-button"> {{ 'app.privacy_policy' | translate }} </a>
-      <a class="mobile-menu-button">
-        {{ 'app.terms_and_conditions' | translate }}
-      </a>
-    </div>
-  </div>
-
-  <div class="menu__footer">
-    <app-donation-button
-      class="menu__footer__button"
-      *ngIf="isDonateAvailable"
-    ></app-donation-button>
-    <mzima-client-button
-      fill="outline"
-      color="custom"
-      (buttonClick)="openShare()"
-      class="menu__footer__button"
-    >
-      {{ 'app.share' | translate }}
-      <mat-icon icon svgIcon="share"></mat-icon>
-    </mzima-client-button>
-  </div>
-</div>
+<app-mobile-menu></app-mobile-menu>
 
 <mat-menu #accountMenu="matMenu" class="account-menu">
-  <button mat-menu-item (click)="openSettings()">{{ 'nav.account_settings' | translate }}</button>
+  <button mat-menu-item (click)="openAccountSettings()">
+    {{ 'nav.account_settings' | translate }}
+  </button>
   <button mat-menu-item (click)="logout()">{{ 'nav.logout' | translate }}</button>
 </mat-menu>
