+++ apps/web-mzima-client/src/app/shared/components/toolbar/toolbar.component.ts
@@ -1,28 +1,12 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
-import { MatDialog } from '@angular/material/dialog';
 import { NavigationEnd, Router } from '@angular/router';
-import { LoginComponent } from '@auth';
-import { CollectionsComponent } from '../index';
-import { EnumGtmEvent, EnumGtmSource } from '@enums';
 import { SiteConfigInterface, UserMenuInterface } from '@models';
 import { UserInterface } from '@mzima-client/sdk';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { TranslateService } from '@ngx-translate/core';
-import {
-  AuthService,
-  BreadcrumbService,
-  BreakpointService,
-  EventBusService,
-  EventType,
-  GtmTrackingService,
-  SessionService,
-} from '@services';
+import { BreadcrumbService, EventBusService, EventType, SessionService } from '@services';
 import { filter, Observable } from 'rxjs';
-import { DonationModalComponent } from '../../../settings';
-import { AccountSettingsModalComponent } from '../account-settings-modal/account-settings-modal.component';
-import { ShareModalComponent } from '../share-modal/share-modal.component';
-import { SupportModalComponent } from '../support-modal/support-modal.component';
+import { NavToolbarService } from '../../helpers/navtoolbar.service';
 
 @UntilDestroy()
 @Component({
@@ -35,7 +19,7 @@ export class ToolbarComponent implements OnInit {
   @Input() selectedLanguage: any;
   private userData$: Observable<UserInterface>;
   public isDesktop$: Observable<boolean>;
-  public isLoggedIn = false;
+  public isLoggedIn: boolean;
   public isDonateAvailable = false;
   public profile: UserInterface;
   public showSearchForm: boolean;
@@ -50,19 +34,15 @@ export class ToolbarComponent implements OnInit {
 
   constructor(
     private session: SessionService,
-    private dialog: MatDialog,
-    private authService: AuthService,
     private router: Router,
     private breadcrumbService: BreadcrumbService,
-    private breakpointService: BreakpointService,
-    private translate: TranslateService,
-    private gtmTracking: GtmTrackingService,
     private eventBusService: EventBusService,
     private location: Location,
+    private navToolbarService: NavToolbarService,
   ) {
-    this.userData$ = this.session.currentUserData$.pipe(untilDestroyed(this));
-    this.isDesktop$ = this.breakpointService.isDesktop$.pipe(untilDestroyed(this));
-    this.siteConfig = this.session.getSiteConfigurations();
+    this.userData$ = this.navToolbarService.getUserData(this);
+    this.isDesktop$ = this.navToolbarService.getScreenSizeAsync(this);
+    this.siteConfig = this.session.getSiteConfigurations(); // Remove this line?
     this.isDonateAvailable = <boolean>this.session.getSiteConfigurations().donation?.enabled;
 
     this.router.events.pipe(filter((event) => event instanceof NavigationEnd)).subscribe(() => {
@@ -86,153 +66,22 @@ export class ToolbarComponent implements OnInit {
     this.userData$.subscribe((userData) => {
       this.profile = userData;
       this.isLoggedIn = !!userData.userId;
-      this.isAdmin = userData.role === 'admin';
-      this.canRegister = !this.siteConfig.private && !this.siteConfig.disable_registration;
-      this.initMenu();
     });
   }
 
-  private initMenu() {
-    this.menu = [
-      {
-        label: 'nav.collections',
-        icon: 'collections',
-        visible: true,
-        action: () => this.openCollections(),
-      },
-      {
-        label: 'nav.help_support',
-        icon: 'info',
-        visible: true,
-        action: () => this.openSupportModal(),
-      },
-      {
-        label: 'nav.my_account',
-        icon: 'account',
-        visible: this.isLoggedIn,
-        action: () => this.openSettings(),
-        separator: true,
-      },
-      {
-        label: 'nav.logout',
-        icon: 'logout',
-        visible: this.isLoggedIn,
-        action: () => this.logout(),
-      },
-      {
-        label: 'nav.login',
-        icon: 'auth',
-        visible: !this.isLoggedIn && !this.canRegister,
-        action: () => this.openLogin(),
-      },
-      {
-        label: 'nav.login_register',
-        icon: 'auth',
-        visible: !this.isLoggedIn && this.canRegister,
-        action: () => this.openLogin(),
-      },
-    ];
-  }
-
-  public showDonation(): void {
-    this.toggleBurgerMenu(false);
-    this.dialog.open(DonationModalComponent, {
-      width: '100%',
-      maxWidth: 564,
-      panelClass: 'modal',
-    });
-  }
-
-  public openSettings(): void {
-    this.toggleBurgerMenu(false);
-    this.dialog.open(AccountSettingsModalComponent, {
-      width: '100%',
-      maxWidth: 800,
-      panelClass: ['modal', 'account-settings-modal'],
-    });
+  public openAccountSettings(): void {
+    this.navToolbarService.openAccountSettings();
   }
 
   public logout(): void {
-    this.authService.logout();
+    this.navToolbarService.logout();
   }
 
-  public openShare() {
-    this.dialog.open(ShareModalComponent, {
-      width: '100%',
-      maxWidth: 564,
-      panelClass: 'modal',
-      data: {
-        title: this.translate.instant(this.pageTitle),
-        description: this.translate.instant(this.pageTitle),
-      },
-    });
-  }
-
-  public toggleBurgerMenu(value?: boolean): void {
-    this.isBurgerMenuOpen = value ?? !this.isBurgerMenuOpen;
-    if (this.isBurgerMenuOpen) {
-      document.body.classList.add('burger-menu-open');
-    } else {
-      document.body.classList.remove('burger-menu-open');
-    }
-  }
-
-  createRouterLink(route: string) {
-    if (route !== 'map' && route !== 'feed') return route;
-    return this.router.url.includes('collection')
-      ? `${route}/collection/${this.router.url.split('/').pop() || ''}`
-      : route;
-  }
-
-  registerPage(event: MouseEvent, router: string, label: string) {
-    label = this.translate.instant(label);
-    event.preventDefault();
-    this.gtmTracking.registerEvent(
-      {
-        event: EnumGtmEvent.PageView,
-        // @ts-ignore
-        source: EnumGtmSource[label],
-      },
-      GtmTrackingService.MapPath(`/${router}`),
-    );
-  }
-
-  private openCollections(): void {
-    this.toggleBurgerMenu(false);
-    const dialogRef = this.dialog.open(CollectionsComponent, {
-      width: '100%',
-      maxWidth: '768px',
-      panelClass: ['modal', 'collections-modal'],
-    });
-
-    dialogRef.afterClosed().subscribe({
-      next: (response) => {
-        response ? console.log(response) : null;
-      },
-    });
-  }
-
-  private openLogin(): void {
-    this.toggleBurgerMenu(false);
-    this.dialog.open(LoginComponent, {
-      width: '100%',
-      maxWidth: 576,
-      panelClass: ['modal', 'login-modal'],
-      data: {
-        isSignupActive: this.canRegister,
-      },
-    });
+  public toggleBurgerMenu(): void {
+    this.navToolbarService.toggleBurgerMenu(); //toggles true & false for hamburger button
   }
 
   public back(): void {
     this.location.back();
-  }
-
-  public openSupportModal(): void {
-    this.dialog.open(SupportModalComponent, {
-      width: '100%',
-      maxWidth: 768,
-      panelClass: ['modal', 'support-modal'],
-    });
   }
 }
